refactor(base-app): rename textAdaptationPreview feature to textSettingsAdaptationPreview
