+++ apps/base-app/src/features/textSettingsAdaptationPreview/lib/visualEngine/adaptHtmlElementAsync.ts
