+++ apps/base-app/src/features/textSettingsAdaptationPreview/ui/TextSettingsAdaptationPreview.vue
